:fire: :wrench: renamed the environment file
